+++ src/main/java/com/example/waiter/Controllers/DishController.java
@@ -1,7 +1,7 @@
 package com.example.waiter.Controllers;
 
 import com.example.waiter.Entities.Dish;
-import com.example.waiter.Repositories.DishRepository;
+import com.example.waiter.Services.DishService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -12,52 +12,31 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.validation.Valid;
-import java.util.Optional;
 
 @Controller
 public class DishController {
     @Autowired
-    DishRepository dishRepository;
+    DishService dishService;
     @PostMapping("/dishSubmit")
     public ModelAndView addDish(@Valid Dish dish, BindingResult bindingResult) {
-        if(bindingResult.hasErrors()){
-            return new ModelAndView("/addDish");
-        } else {
-            dishRepository.save(dish);
-            return new ModelAndView("redirect:/homePageWaiter");
-        }
+        return dishService.addDishSubmit(dish, bindingResult);
     }
 
     @GetMapping("/addDish")
     public String addDish(Model model) {
-        model.addAttribute("dish", new Dish());
-        return ("/addDish");
+        return dishService.addDish(model);
     }
     @GetMapping("/editDish/{dishId}")
     public String editDish(@PathVariable(name = "dishId") Long dishId, Model model) {
-        Optional<Dish> optionalDish = dishRepository.findById(dishId);
-        if (optionalDish.isPresent()) {
-            model.addAttribute("dish", optionalDish.get());
-        } else {
-            model.addAttribute("dish", "Error!");
-            model.addAttribute("errorMsg", " Not existing dish with id = " + dishId);
-        }
-        return "/editDish";
+        return dishService.editDish(dishId, model);
     }
 
     @PostMapping("/updateDish")
     public ModelAndView updateDish(@Valid Dish dish, BindingResult bindingResult, Model model) {
-        if (bindingResult.hasErrors()) {
-            ;
-            return new ModelAndView("/editDish");
-        } else {
-            dishRepository.save(dish);
-            return new ModelAndView("redirect:/restaurantMenu");
-        }
+        return dishService.updateDish(dish, bindingResult, model);
     }
     @PostMapping("/deleteDish/{dishId}")
     public ModelAndView deleteDish(@PathVariable(name = "dishId") Long dishId) {
-        dishRepository.deleteById(dishId);
-        return new ModelAndView("redirect:/restaurantMenu");
+        return dishService.deleteDish(dishId);
     }
 }
