+++ src/main/java/com/example/waiter/Controllers/DrinkController.java
@@ -1,7 +1,7 @@
 package com.example.waiter.Controllers;
 
 import com.example.waiter.Entities.Drink;
-import com.example.waiter.Repositories.DrinkRepository;
+import com.example.waiter.Services.DrinkService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -12,51 +12,31 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.validation.Valid;
-import java.util.Optional;
 
 @Controller
 public class DrinkController {
     @Autowired
-    DrinkRepository drinkRepository;
+    DrinkService drinkService;
     @PostMapping("/drinkSubmit")
     public ModelAndView addDrink(@Valid Drink drink, BindingResult bindingResult) {
-        if(bindingResult.hasErrors()){
-            return new ModelAndView("/addDrink");
-        } else {
-            drinkRepository.save(drink);
-            return new ModelAndView("redirect:/homePageWaiter");
-        }
+        return drinkService.addDrink(drink, bindingResult);
     }
 
     @GetMapping("/addDrink")
     public String addDrink(Model model) {
-        model.addAttribute("drink", new Drink());
-        return ("/addDrink");
+        return drinkService.addDrink(model);
     }
     @GetMapping("/editDrink/{drinkId}")
     public String editDrink(@PathVariable(name = "drinkId") Long drinkId, Model model) {
-        Optional<Drink> optionalDrink = drinkRepository.findById(drinkId);
-        if (optionalDrink.isPresent()) {
-            model.addAttribute("drink", optionalDrink.get());
-        } else {
-            model.addAttribute("drink", "Error!");
-            model.addAttribute("errorMsg", " Not existing drink with id = " + drinkId);
-        }
-        return "/editDrink";
+        return drinkService.editDrink(drinkId, model);
     }
 
     @PostMapping("/updateDrink")
     public ModelAndView updateDrink(@Valid Drink drink, BindingResult bindingResult, Model model) {
-        if (bindingResult.hasErrors()) {
-            return new ModelAndView("/editDrink");
-        } else {
-            drinkRepository.save(drink);
-            return new ModelAndView("redirect:/restaurantMenu");
-        }
+        return drinkService.updateDrink(drink, bindingResult, model);
     }
     @PostMapping("/deleteDrink/{drinkId}")
     public ModelAndView deleteDrink(@PathVariable(name = "drinkId") Long drinkId) {
-        drinkRepository.deleteById(drinkId);
-        return new ModelAndView("redirect:/restaurantMenu");
+        return drinkService.deleteDrink(drinkId);
     }
 }
