+++ src/main/java/com/example/waiter/Controllers/MenuController.java
@@ -1,7 +1,5 @@
 package com.example.waiter.Controllers;
 
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
