+++ src/main/java/com/example/waiter/Controllers/RestaurantMenu.java
@@ -1,9 +1,6 @@
 package com.example.waiter.Controllers;
 
-import com.example.waiter.Entities.Dish;
-import com.example.waiter.Entities.Drink;
-import com.example.waiter.Repositories.DishRepository;
-import com.example.waiter.Repositories.DrinkRepository;
+import com.example.waiter.Services.RestaurantMenuService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -11,15 +8,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 @Controller
 public class RestaurantMenu {
     @Autowired
-    DishRepository dishRepository;
-    @Autowired
-    DrinkRepository drinkRepository;
+    private RestaurantMenuService restaurantMenuService;
     @GetMapping("/restaurantMenu")
     public String getAllDishes(Model model){
-        Iterable<Dish> allDishes = dishRepository.findAll();
-        model.addAttribute("allDishes", allDishes);
-        Iterable<Drink> allDrinks = drinkRepository.findAll();
-        model.addAttribute("allDrinks", allDrinks);
-        return "/restaurantMenu";
+        return restaurantMenuService.getAllDishes(model);
     }
 }
