+++ src/main/resources/templates/activeOrders.html
@@ -38,7 +38,7 @@
 <div class = "fr">
   <table class="table table-bordered">
     <thead>
-    <tr class="text-center">-->
+    <tr class="text-center">
       <td colspan="6" style="align:center"><h1>ACTIVE ORDERS</h1></td>
     </tr>
     <tr>
@@ -58,7 +58,7 @@
       <td th:text="${order.totalPrice}"></td>
       <td th:text="${order.orderDate}"></td>
       <td>
-        <a th:href="@{editOrder/{orderId}(orderId=${order.id})}">Update</a>
+        <a th:href="@{editOrder/{orderId}(orderId=${order.id})}">Update status</a>
       </td>
     </tr>
     </tbody>
