+++ src/main/resources/templates/editDrink.html
@@ -13,7 +13,6 @@
     <title>Edit Drink</title>
 </head>
 <body>
-<h1 align="center">Edit Drink</h1>
 <span th:if="${drink == 'Error!'}">Error!-->
         <span th:text="${errorMsg}"></span>
       </span>
